Added DCBM generator and script. Separated random network generation out into separate module.
